Hide social media image on smaller screens
Social media image in the footer was taking significant space on smaller screens due to its size and proportion, which was impairing user navigation and overall experience, especially on mobile devices. To solve this, modified Bootstrap responsive classes in index.html to hide this image on XS screens; it will still be visible on screens that are MD or larger.